requirements.txt
apache-beam==2.38.0
google-cloud-storage==1.44.0
pandas==1.3.5
pytest==7.1.2
